[PyTorch] Rationalize test_grouped_linear.py structure
Structural changes:
- Organize into three labeled sections: te.GroupedLinear (module API),
  raw GEMM kernels (cpp_extensions), te.ops.GroupedLinear (ops/fuser API)
- Move fused-path helpers and tests (_reset_fp8_state, _run_grouped_linear_path,
  test_grouped_linear_grouped_tensor_path_*, test_grouped_linear_fused_path_*)
  before the GEMM section so all te.GroupedLinear tests are contiguous

Fold redundant tests:
- test_grouped_linear_accuracy_single_gemm → num_gemms=[1,3,6]
- test_padding_grouped_linear_accuracy_save_original_input →
  save_original_input parametrize axis on test_padding_grouped_linear_accuracy
- test_grouped_linear_grouped_tensor_path_single_grouped_bias_delay_wgrad →
  single_grouped_bias parametrize axis on test_grouped_linear_grouped_tensor_path_matches_legacy

Fix test_grouped_mlp reference precision and tolerances:
- Replace float32 reference tensors with make_reference_and_test_tensors
  (float64 CPU reference, target dtype CUDA test)
- Replace loose rtol=0.125/0.25 with dtype_tols()/quantization_tols()

Promote shared helpers to utils.py:
- maybe_skip_quantization: was duplicated in test_fusible_ops.py and
  test_grouped_linear.py; now in utils.py
- make_reference_and_test_tensors: same
- dtype_tols: was duplicated locally in test_grouped_linear.py; now
  imported from utils.py (where it already existed)
